+++ ichnaea/api/submit/views.py
@@ -39,7 +39,7 @@ class BaseSubmitView(BaseAPIView):
 
     def emit_upload_metrics(self, value, api_key):
         tags = None
-        if api_key.should_log('submit'):
+        if api_key.valid_key:
             tags = ['key:%s' % api_key.valid_key]
         self.stats_client.incr('data.batch.upload', tags=tags)