+++ ichnaea/models/api.py
@@ -41,9 +41,6 @@ class ApiKey(_Model):
 
     valid_key = Column(String(40), primary_key=True)  #: UUID API key.
     maxreq = Column(Integer)  #: Maximum number of requests per day.
-    log_locate = Column(Boolean)  # unused
-    log_region = Column(Boolean)  # unused
-    log_submit = Column(Boolean)  # unused
     allow_fallback = Column(Boolean)  #: Use the fallback source?
     allow_locate = Column(Boolean)  #: Allow locate queries?
     shortname = Column(String(40))  #: Short descriptive name.
@@ -87,13 +84,6 @@ class ApiKey(_Model):
                     self.fallback_url and
                     self.fallback_ratelimit is not None and
                     self.fallback_ratelimit_interval)
-
-    def should_log(self, api_type):
-        # This used to differentiate between basic/extended logging
-        # for each API type (locate, region, submit).
-        if self.valid_key:
-            return True
-        return False
 
     def __str__(self):
         return '<ApiKey>: %s' % self.valid_key